requirements.txt
streamlit
openai
pydub
python-dotenv
pandas
rapidfuzz
